[PGPRO-11599] Fix wrong results returned when order_by_attach=TRUE.
With order_by_attach=TRUE, the RUM index arranges the entries in
the posting lists (for keys that have additional information) in
order of sorting by additional information. For the remaining keys,
the records are sorted by tid. The scanGetItemRegular() function
assumes that all posting lists have the same sorting, so it was
not suitable for sorting in this case.

In order to fix this, the scanGetItemRegularAltorderWithAddInfo()
function and tests have been added to verify that the index works
correctly in the described situation.

Tags: rum